requirements.txt
# ConsultantPro - Python Dependencies
# Phase 1: Core Skeleton

# Django Core (4.2+)
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.27.0

# AWS S3 (for Documents module)
boto3==1.34.11
django-storages==1.14.2

# Stripe (for Finance module)
stripe==7.9.0

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0

# Production Server (optional, for deployment)
gunicorn==21.2.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.0.0
